requirements.txt
absl-py==2.1.0
accelerate==0.29.2
addict==2.4.0
aiofiles==23.2.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
av==12.3.0
bcrypt==4.2.0
beartype==0.18.5
beautifulsoup4==4.12.3
bitsandbytes==0.43.2
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colossalai==0.4.0
comm==0.2.2
contexttimer==0.3.3
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
Deprecated==1.2.14
diffusers==0.27.2
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.8.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
fabric==3.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
future==1.0.0
galore-torch==1.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
gradio==4.39.0
gradio_client==1.1.1
grpcio==1.65.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.1
identify==2.6.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
invoke==2.2.0
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lightning-utilities==0.11.7
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmengine==0.10.4
mpmath==1.3.0
msgpack==1.0.8
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nodeenv==1.9.1
numpy==1.26.4
openai==1.37.0
opencv-python-headless==4.10.0.84
orjson==3.10.6
packaging==24.1
pandarallel==1.6.5
pandas==2.2.2
paramiko==3.4.0
parso==0.8.4
peft==0.12.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plumbum==1.8.3
pre-commit==3.7.1
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
ray==2.32.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rotary-embedding-torch==0.5.3
rpds-py==0.19.1
rpyc==6.0.0
ruff==0.5.4
safetensors==0.4.3
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.11.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spaces==0.29.0
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensornvme==0.1.0
termcolor==2.4.0
timm==0.9.16
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
torchmetrics==1.4.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.39.3
triton==2.2.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.26.3
wandb==0.17.5
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
yapf==0.40.2
zipp==3.19.2